requirements.txt
transformers
pillow
streamlit
streamlit-webrtc
opencv-python
mediapipe
numpy
torch
torchvision
